change all floats to double and add traits for router

+++ rust_qsim/src/simulation/config.rs
@@ -515,7 +515,7 @@ pub struct DrtService {
 pub struct Simulation {
     pub start_time: u32,
     pub end_time: u32,
-    pub sample_size: f32,
+    pub sample_size: f64,
     pub stuck_threshold: u32,
     pub main_modes: Vec<String>,
 }
